+++ phy_80211.cc
@@ -999,9 +999,7 @@ kis_80211_phy::kis_80211_phy(int in_phyid) :
 
                     auto pcapng = 
 					std::make_shared<pcapng_stream_packetchain<pcapng_phy80211_accept_ftor, pcapng_stream_select_ftor>>(&con->response_stream(),
-							pcapng_phy80211_accept_ftor(mac),
-							pcapng_stream_select_ftor(),
-                            1024*512);
+							pcapng_phy80211_accept_ftor(mac), pcapng_stream_select_ftor(), (size_t) 1024*512);
         
                     con->clear_timeout();
                     con->set_target_file(fmt::format("kismet-80211-bssid-{}.pcapng", mac));
